+++ chevah/compat/tests/elevated/__init__.py
@@ -6,12 +6,10 @@ functions.
 """
 
 from chevah.compat import process_capabilities
-from chevah.empirical.testcase import (
-    ChevahTestCase,
-    setup_os,
-    teardown_os,
-    )
 from chevah.compat.testing import (
+    CompatTestCase,
+    setup_access_control,
+    teardown_access_control,
     TEST_GROUPS,
     TEST_USERS,
     )
@@ -33,18 +31,18 @@ def runElevatedTest():
 def setup_package():
     # Don't run these tests if we can not access privileged OS part.
     if not runElevatedTest():
-        raise ChevahTestCase.skipTest()
+        raise CompatTestCase.skipTest()
     # Initialize the testing OS.
 
     try:
-        setup_os(users=TEST_USERS, groups=TEST_GROUPS)
+        setup_access_control(users=TEST_USERS, groups=TEST_GROUPS)
     except:
         import traceback
         print traceback.format_exc()
         print "Failed to initialized the system accounts!"
-        teardown_os(users=TEST_USERS, groups=TEST_GROUPS)
+        teardown_access_control(users=TEST_USERS, groups=TEST_GROUPS)
         raise
 
 
 def teardown_package():
-    teardown_os(users=TEST_USERS, groups=TEST_GROUPS)
+    teardown_access_control(users=TEST_USERS, groups=TEST_GROUPS)