+++ paver.sh
@@ -4,6 +4,28 @@
 #
 # Helper script for bootstraping the build system on Unix/Msys.
 # It will write the default values into 'DEFAULT_VALUES' file.
+#
+# To use this script you will need to publish binary archive files for the
+# following components:
+#
+# * Python main distribution
+# * pip
+# * setuptools
+#
+
+# Script initialization.
+set -o nounset
+set -o errexit
+set -o pipefail
+
+# Initialize default value.
+COMMAND=${1-''}
+DEBUG=${DEBUG-0}
+
+BINARY_DIST_URI="http://binary.chevah.com/production/python"
+PAVER_VERSION='1.2.1'
+PIP_VERSION="1.4.1-c1"
+SETUPTOOLS_VERSION="1.4.1"
 
 # Set default locale.
 # We use C (alias for POSIX) for having a basic default value and
@@ -25,11 +47,11 @@ WAS_PYTHON_JUST_INSTALLED=0
 DIST_FOLDER='dist'
 
 # Path global variables.
-PROJECT_ROOT=""
-BOOTSTRAP_PATH=""
 BUILD_FOLDER=""
+CACHE_FOLDER="cache"
 PYTHON_BIN=""
 PYTHON_LIB=""
+LOCAL_PYTHON_BINARY_DIST=""
 
 
 clean_build() {
@@ -47,14 +69,28 @@ clean_build() {
 
 
 #
+# Wrapper for executing a command and exiting on failure.
+#
+execute() {
+    if [ $DEBUG -ne 0 ]; then
+        echo "Executing:" $@
+    fi
+
+    #Make sure $@ is called in quotes as otherwise it will not work.
+    set +e
+    "$@"
+    exit_code=$?
+    set -e
+    if [ $exit_code -ne 0 ]; then
+        echo "Fail:" $@
+        exit 1
+    fi
+}
+
+#
 # Update global variables with current paths.
 #
 update_path_variables() {
-
-    # Initialize PROJECT_ROOT and later fix it.
-    PROJECT_ROOT=`pwd`
-    PROJECT_ROOT=${PROJECT_ROOT%chevah*}
-    PROJECT_ROOT=${PROJECT_ROOT}chevah
 
     if [ "${OS}" = "windows" ] ; then
         PYTHON_BIN="/lib/python.exe"
@@ -64,11 +100,11 @@ update_path_variables() {
         PYTHON_LIB="/lib/${PYTHON_VERSION}/"
     fi
 
-    BOOTSTRAP_PATH=${PROJECT_ROOT}/brink
-
     BUILD_FOLDER="build-${OS}-${ARCH}"
     PYTHON_BIN="${BUILD_FOLDER}${PYTHON_BIN}"
     PYTHON_LIB="${BUILD_FOLDER}${PYTHON_LIB}"
+
+    LOCAL_PYTHON_BINARY_DIST="$PYTHON_VERSION-$OS-$ARCH"
 
     export PYTHONPATH=${BUILD_FOLDER}
 }
@@ -84,17 +120,6 @@ write_default_values() {
 #
 install_brink() {
     raw_version=`grep "BRINK_VERSION =" pavement.py`
-    exit_code=$?
-
-    # This code will be deprecated in future versions since we will no longer
-    # support pavement_lib imports.
-    if [ $exit_code -ne 0 ]; then
-        # Brink version was not found, so we go with default.
-        echo "Installing latest version of pavement_lib.py..."
-        PAVEMENT_LIB_PATH=${BOOTSTRAP_PATH}/pavement_commons.py
-        cp ${PAVEMENT_LIB_PATH} pavement_lib.py
-        return
-    fi
 
     # Extract version and remove quotes.
     version=${raw_version#BRINK_VERSION = }
@@ -110,43 +135,111 @@ install_brink() {
 
     echo "Installing version: chevah-brink==$version of brink..."
 
+    pip install "chevah-brink==$version"
+}
+
+
+#
+# Wrapper for python pip command.
+# * $1 - command name
+# * $2 - package_name and optional version.
+#
+pip() {
+    set +e
     ${PYTHON_BIN} -m \
-        pip.__init__ install chevah-brink==$version \
+        pip.__init__ $1 $2 \
             --index-url=http://172.20.0.1:10042/simple \
-            --download-cache=${PROJECT_ROOT}/brink/cache/pypi \
-            --find-links=file://${PROJECT_ROOT}/brink/cache/pypi \
+            --download-cache=${CACHE_FOLDER}/pypi \
+            --find-links=file://${CACHE_FOLDER}/cache/pypi \
             --upgrade
 
     exit_code=$?
+    set -e
     if [ $exit_code -ne 0 ]; then
         echo "Failed to install brink."
         exit 1
     fi
 }
 
-get_python_version() {
-    raw_version=`grep "PYTHON_VERSION =" pavement.py`
+
+#
+# Download and extract a binary distribution.
+#
+get_binary_dist() {
+    dist_name=$1
+
+    echo "Getting $dist_name ..."
+
+    tar_gz_file=${dist_name}.tar.gz
+    tar_file=${dist_name}.tar
+
+    mkdir -p ${CACHE_FOLDER}
+    pushd ${CACHE_FOLDER}
+
+        # Get and extract archive.
+        rm -rf $dist_name
+        rm -f $tar_gz_file
+        rm -f $tar_file
+        # Use 1M dot to reduce console polution.
+        execute wget --progress=dot -e dotbytes=1M $BINARY_DIST_URI/${tar_gz_file}
+        execute gunzip $tar_gz_file
+        execute tar -xf $tar_file
+        rm -f $tar_gz_file
+        rm -f $tar_file
+
+    popd
+}
+
+#
+# Return a version defined in pavement.py file.
+#
+get_version_from_pavement() {
+    local name
+    local raw_version
+    local version=''
+
+    name=$1
+
+    set +e
+    raw_version=`grep "$name =" pavement.py`
     exit_code=$?
+    set -e
     if [ $exit_code -ne 0 ]; then
-        # Python version was not found, so we go with default.
-        return
-    fi 
+        # Version was not found, so we go with default.
+        return version
+    fi
 
     # Extract version and remove quotes.
-    version=${raw_version#PYTHON_VERSION = }
+    version=${raw_version#$name = }
     version=${version#\'}
     version=${version%\'}
     version=${version#\"}
     version=${version%\"}
-    PYTHON_VERSION="python$version"
+
+    echo $version
 }
+
+
+#
+# Return the version of python used in the current repository.
+#
+get_python_version() {
+    local version
+
+    version=$(get_version_from_pavement PYTHON_VERSION)
+
+    echo "python$version"
+}
+
 
 #
 # Copy python to build folder from binary distribution.
 #
 copy_python() {
 
-    PYTHON_DISTRIBUTABLE=${BOOTSTRAP_PATH}/cache/${PYTHON_VERSION}-${OS}-${ARCH}
+    local python_distributable="${CACHE_FOLDER}/${PYTHON_VERSION}-${OS}-${ARCH}"
+    local pip_package="pip-$PIP_VERSION"
+    local setuptools_package="setuptools-$SETUPTOOLS_VERSION"
 
     # Check that python dist was installed
     if [ ! -s ${PYTHON_BIN} ]; then
@@ -155,37 +248,29 @@ copy_python() {
         mkdir -p ${BUILD_FOLDER}
 
         # If we don't have a cached python distributable,
-        # get one.
-        if [ ! -d ${PYTHON_DISTRIBUTABLE} ]; then
+        # get one together with default build system.
+        if [ ! -d ${python_distributable} ]; then
             echo "No ${PYTHON_VERSION} environment. Start downloading it..."
-            pushd ${BOOTSTRAP_PATH}
-            git pull
-            ./make-it-happen.sh get_python ${PYTHON_VERSION} ${OS}-${ARCH}
-            popd
+            get_binary_dist ${PYTHON_VERSION}-${OS}-${ARCH}
+            get_binary_dist "$pip_package"
+            get_binary_dist "$setuptools_package"
         fi
 
-        cp -R ${PYTHON_DISTRIBUTABLE}/* ${BUILD_FOLDER}
-        cp -r ${BUILD_FOLDER}/lib/config/include ${BUILD_FOLDER}/
-        # Copy pywintypes25.dll as it is required by paver on windows.
-        if [ "$OS" = "windows" ]; then
-            cp -R ${PYTHON_DISTRIBUTABLE}/lib/pywintypes25.dll .
-        fi
+        echo "Copying bootstraping files... "
+        cp -R ${python_distributable}/* ${BUILD_FOLDER}
+
+        cp -RL "${CACHE_FOLDER}/$pip_package/pip" ${PYTHON_LIB}/site-packages/
+        cp -RL "${CACHE_FOLDER}/$setuptools_package/setuptools" ${PYTHON_LIB}/site-packages/
+        cp -RL "${CACHE_FOLDER}/$setuptools_package//setuptools.egg-info" ${PYTHON_LIB}/site-packages/
+        cp "${CACHE_FOLDER}/$setuptools_package/pkg_resources.py" ${PYTHON_LIB}/site-packages/
+        cp "${CACHE_FOLDER}/$setuptools_package/easy_install.py" ${PYTHON_LIB}/site-package
+
+        # Once we have pip, we can use it.
+        pip install "paver==$PAVER_VERSION"
 
         WAS_PYTHON_JUST_INSTALLED=1
     fi
-}
 
-
-# Copy base python packages to Python.
-# We can not install them using pip, since we don't have pip or easy_install.
-copy_base_packages() {
-    # Always update paver ... at least until we have a stable buildsystem.
-    cp -RL ${BOOTSTRAP_PATH}/paver/paver ${PYTHON_LIB}/site-packages/
-    cp -RL ${BOOTSTRAP_PATH}/pip/pip ${PYTHON_LIB}/site-packages/
-    cp -RL ${BOOTSTRAP_PATH}/distribute/setuptools ${PYTHON_LIB}/site-packages/
-    cp -RL ${BOOTSTRAP_PATH}/distribute/distribute.egg-info ${PYTHON_LIB}/site-packages/
-    cp ${BOOTSTRAP_PATH}/distribute/pkg_resources.py ${PYTHON_LIB}/site-packages/
-    cp ${BOOTSTRAP_PATH}/distribute/easy_install.py ${PYTHON_LIB}/site-packages/
 }
 
 
@@ -200,9 +285,11 @@ install_dependencies(){
 
     install_brink
 
+    set +e
     ${PYTHON_BIN} -c 'from paver.tasks import main; main()' deps
-    python_exit_code=$?
-    if [ $python_exit_code -ne 0 ]; then
+    exit_code=$?
+    set -e
+    if [ $exit_code -ne 0 ]; then
         echo 'Failed to run the inital "paver deps" command.'
         exit 1
     fi
@@ -238,7 +325,7 @@ detect_os() {
     elif [ "${OS}" = "sunos" ] ; then
 
         OS="solaris"
-        ARCH=`uname -p`	
+        ARCH=`uname -p`
         VERSION=`uname -r`
 
         if [ "$ARCH" = "i386" ] ; then
@@ -310,7 +397,7 @@ detect_os() {
                     OS='ubuntu1204'
                 ;;
                 # Lie for dumol's Gentoo. Separate so that it's clear
-                '2.1')
+                '2.2')
                     OS='ubuntu1204'
                 ;;
                 *)
@@ -318,7 +405,7 @@ detect_os() {
                     exit 1
                 ;;
             esac
-    	    
+
         elif [ -f /etc/slackware-version ] ; then
 
             # For Slackware, for now we use Ubuntu 10.04.
@@ -339,7 +426,7 @@ detect_os() {
     		echo 'Unsuported OS X version.'
     		exit 1
     	fi
-    	
+
     	osx_arch=`uname -m`
     	if [ "$osx_arch" = "Power Macintosh" ] ; then
     		ARCH='ppc'
@@ -369,36 +456,41 @@ detect_os() {
 # Put default values and create them as global variables.
 OS='not-detected-yet'
 ARCH='x86'
-PYTHON_VERSION="python2.5"
+PYTHON_VERSION=$(get_python_version)
 
 detect_os
-get_python_version
 update_path_variables
 
-if [ "$1" = "clean" ] ; then
+if [ "$COMMAND" = "clean" ] ; then
     clean_build
     exit 0
 fi
 
-if [ "$1" = "get_default_values" ] ; then
+if [ "$COMMAND" = "get_default_values" ] ; then
     write_default_values
+    exit 0
+fi
+
+if [ "$COMMAND" = "get_python" ] ; then
+    get_binary_dist $2
     exit 0
 fi
 
 check_source_folder
 write_default_values
 copy_python
-copy_base_packages
 install_dependencies
 
 # Always update brink when running buildbot tasks.
 for paver_task in "deps" "test_os_dependent" "test_os_independent"; do
-    if [ "$1" == "$paver_task" ] ; then
+    if [ "$COMMAND" == "$paver_task" ] ; then
         install_brink
     fi
 done
 
 # Now that we have Python and Paver, let's call Paver from Python :)
-${PYTHON_BIN} -c 'from paver.tasks import main; main()' $@
-python_exit_code=$?
-exit $python_exit_code
+set +e
+${PYTHON_BIN} -c 'from paver.tasks import main; main()' "$@"
+exit_code=$?
+set -e
+exit $exit_code